backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
Django==5.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
